Error range changed peptides that contain Q or N
The default threshold (-03 to +1.3) was changed to -0.3 to +0.3 (the accuracy interval of the MALDI-ToF-MS), except in Q or N containing peptides: COL1A2 793–816 (D), COL1A2 502–519 (C), COL1A1 586–618 (F), COL1A1 586–618+16 (F'), COL1A1 508–519 (P1) where the larger accuracy interval was kept (-0.3 to +1.3). See the comments about this change in the transform function.